requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
Django==4.1.4
django-filter==22.1
django-plpy==0.1.0
django-searchable-select==1.5.0
lxml==4.9.1
numpy==1.23.3
pandas==1.5.0
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
PyPDF2==2.11.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tabula-py==2.5.1
